[level 1] Title: 3진법 뒤집기, Time: 0.19 ms, Memory: 71.1 MB -BaekjoonHub

+++ 프로그래머스/lv1/68935. 3진법 뒤집기/3진법 뒤집기.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int solution(long n) {
+    public int solution(int n) {
         int answer = 0;
         
         long ternary = 0;
